requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Session==0.5.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
openpyxl==3.1.2
pycparser==2.21
pymon==1.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.30.0
six==1.16.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0